pyproject.toml
[project]
name = "tarmac"
version = "0.1.0"
description = "Reliable deployments, repeatable workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotmap>=1.3.30",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "uv>=0.6.10",
]

[project.scripts]
tarmac = "tarmac.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.6",
]
